fitur untuk merubah roles dan protect fungsi delete untuk role member

+++ convex/files.ts
@@ -33,7 +33,8 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
   }
   // variable hasAccess mengecek apakah user yg berhasil di query terdapat orgId dan tokenIdentifier yang sama dengan yg dikirim dari client
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
   if (!hasAccess) {
     return null;
   }
@@ -148,6 +149,15 @@ export const deleteFile = mutation({
     if (!access) {
       throw new ConvexError("You do not have access to this file");
     }
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("You do not have admin access to delete this file");
+    }
+
     // jika berhasil, jalankan proses delete file
     await ctx.db.delete(args.fileId);
   },
